requirements.txt
beepy
Pillow
numpy